refactor: RoomCreate Screen Content To FirstScreen
